+++ octue/__init__.py
@@ -1,9 +1,8 @@
 import logging
 
-from .cli import octue_cli
 from .logging_handlers import LOG_FORMAT
 from .runner import Runner
 
 
-__all__ = "LOG_FORMAT", "octue_cli", "Runner"
+__all__ = "LOG_FORMAT", "Runner"
 package_logger = logging.getLogger(__name__)
