--- octue/resources/communication/__init__.py
@@ -1,4 +1,0 @@
-from .google_pub_sub.service import Service
-
-
-__all__ = ("Service",)
