+++ tests/deployment/test_cloud_run.py
@@ -3,9 +3,9 @@ import json
 import uuid
 from unittest import TestCase, mock
 
+from octue.cloud.pub_sub.service import Service
 from octue.deployment.google import cloud_run
-from octue.resources.communication.google_pub_sub.service import Service
-from octue.resources.communication.service_backends import GCPPubSubBackend
+from octue.resources.service_backends import GCPPubSubBackend
 from tests import TEST_PROJECT_NAME
